refactor(device): Cameras — modèles Eloquent vers Domain/Models (Closes #6842) ; backups fail-loud (Part of #6836)
Lot PM principal 2026-09-05 (issues ouvertes) :
- #6842 (BC-19 DEVICE) : api/app/Modules/Cameras/Domain/{Camera,CameraAccessLog,
  CameraAccessToken,CameraPermission}.php -> Domain/Models/ (namespace
  App\Modules\Cameras\Domain\Models) ; 14 fichiers référents mis à jour
  (Actions, Controllers, Services, Policy, AuthServiceProvider, 3 tests).
  Vérifié : 0 référence restante à l'ancien chemin. Gate PHPStan en CI.
- #6836 (P1 data-loss) : database-backup.yml — plus de no-op silencieux quand
  les secrets manquent : ::error + exit 1 (backup quotidien ET drill de
  restauration). La pose des secrets reste une action fondateur (commentée
  dans le workflow).
