+++ api/app/Modules/Cameras/Domain/Models/CameraAccessLog.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Modules\Cameras\Domain;
+namespace App\Modules\Cameras\Domain\Models;
 
 use App\Core\Auth\Domain\Models\Employee;
 use App\Shared\Traits\BelongsToCompany;
