+++ api/app/Modules/Cameras/Infrastructure/Services/CameraService.php
@@ -7,9 +7,9 @@ namespace App\Modules\Cameras\Infrastructure\Services;
 use App\Core\Auth\Domain\Models\Employee;
 use App\Core\Tenant\Domain\Models\Company;
 use App\Exceptions\DomainException;
-use App\Modules\Cameras\Domain\Camera;
-use App\Modules\Cameras\Domain\CameraAccessLog;
-use App\Modules\Cameras\Domain\CameraAccessToken;
+use App\Modules\Cameras\Domain\Models\Camera;
+use App\Modules\Cameras\Domain\Models\CameraAccessLog;
+use App\Modules\Cameras\Domain\Models\CameraAccessToken;
 use App\Modules\Cameras\Infrastructure\Streaming\CameraStreamTokenService;
 use App\Rules\NotPrivateUrl;
 use Illuminate\Support\Carbon;
